Resolved Issue50. Moved display/update code to TimePanel and renamed classes
